requirements.txt
configparser==5.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
